+++ Src/AssemblyInfo.vb
@@ -12,4 +12,4 @@ Imports System.Runtime.InteropServices
 <Assembly: AssemblyCopyright("Copyright © 2026")>
 <Assembly: AssemblyTrademark("")>
 
-<Assembly: AssemblyVersion("1.1.0.*")>
+<Assembly: AssemblyVersion("1.1.1.*")>