+++ Changelog.md
@@ -2,7 +2,13 @@
 
 Toutes les modifications notables apportées au projet VBFileFind sont documentées dans ce fichier.
 
+## [1.11] - 2026-07-31
+### Corrigé
+- Recherche avec l'un ou l'autre encodage (Les 2 encodages) : corrigé depuis la modif. du 16/04/2023 pour détecter l'encodage
+- Correction des recherches insensibles à la casse pour l'option Détecter l'encodage
+
 ## [1.10] - 2026-03-08
+### Corrigé
 - Windows 11 : on peut encore ouvrir les fichiers via le bloc-notes, mais sans recherche de l'occurrence trouvée (ne pas ouvrir l'appli. associée)
 
 ## [1.09] - 2024-08-03